Advection and pressure, with energy output
This is (almost entirely) untested. Code has been added to compute the
nonlinear advection terms, and the pressure terms. Code has also been
added to output the total kinetic energy of the flow at each timestep.
The only testing that has been performed is to set up two flows of the
form (sin(2\pi y), 0) and (0, sin(2\pi x)), and confirm that the energy
remains constant (which is definitely should since there is no
dissipation) and equal to 0.5 (I've not checked if this is the value
that I should expect.