Added velocity output function
Added velocity output function to allow me to examine my simple flow
cases. Currently it looks like I have diagonal striping, which cannot
be correct! Also, the outputVelocity function is getting called only
once when I intend it to be called multiple time - not sure why this
is the case.